adiciona jwt authentication

+++ server/src/routes/memories.ts
@@ -2,9 +2,18 @@ import { FastifyInstance } from 'fastify'
 import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
+const CODE_UNAUTHORIZED = 401
+
 export async function memoriesRoute(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.get('/memories', async (request) => {
+    app.addHook('preHandler', async () => {
+      await request.jwtVerify()
+    })
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -17,15 +26,21 @@ export async function memoriesRoute(app: FastifyInstance) {
     }))
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
 
-    return await prisma.memory.findFirstOrThrow({
+    const memory = await prisma.memory.findFirstOrThrow({
       where: { id },
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(CODE_UNAUTHORIZED).send()
+    }
+
+    return memory
   })
 
   app.post('/memories', async (request) => {
@@ -41,12 +56,12 @@ export async function memoriesRoute(app: FastifyInstance) {
         content,
         isPublic,
         coverUrl,
-        userId: 'a643a5f6-f5fd-11ed-b67e-0242ac120002',
+        userId: request.user.sub,
       },
     })
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -57,6 +72,14 @@ export async function memoriesRoute(app: FastifyInstance) {
     })
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(CODE_UNAUTHORIZED).send()
+    }
 
     return await prisma.memory.update({
       where: {
@@ -70,11 +93,19 @@ export async function memoriesRoute(app: FastifyInstance) {
     })
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(CODE_UNAUTHORIZED).send()
+    }
 
     await prisma.memory.delete({
       where: {
